fix(security): R240 adversarial audit — Sprint 4 P3 sweep (24 files, 28 findings)
#[must_use] annotations on security-critical functions:
- verified_core: compute_verdict, compute_single_verdict
- adaptive_rate: AdaptiveRateLimiter::check()
- circuit_breaker: CircuitBreakerManager::can_proceed()
- cascading: CascadingTracker::check_pipeline()
- behavioral: BehavioralTracker::check_session()
- types/core: has_dangerous_chars()

u64-to-usize content_length cast fixes (compare in u64 space):
- proxy/helpers, smart_fallback, fallback, grpc/upstream
- oauth JWKS, federation JWKS

DPoP u64-to-i64 safe casts (reject > i64::MAX):
- oauth: iat, exp, nbf fields via i64::try_from with fail-closed

Server timestamp safe casts:
- dashboard: started_at i64::try_from instead of `as i64`
- iam: M2M now_secs fail-closed on pre-epoch clock
- policy_lifecycle: staged_secs/now_secs via u64::try_from

Lock poisoning logging (was silent unwrap_or(0)):
- context_isolation: entry_count, session_count
- session_isolator: session_count
- sanitizer: mapping_count
- discovery/guard: current() — parity with check()

Silent .ok()/.unwrap_or_default() observability:
- injection scanner: log AhoCorasick compilation failure
- audit/query/file: log entry serialization failure
- audit/otlp: log hex decode failure in trace ID conversion
- config/manifest: log annotation serialization failure
- discovery/serialize: log clock anomaly in snapshot epoch

+++ vellaveto-http-proxy/src/federation.rs
@@ -502,8 +502,9 @@ impl FederationResolver {
 
         // SECURITY (FIND-R50-001): Check Content-Length header BEFORE downloading
         // to prevent OOM from malicious JWKS endpoints serving gigabytes.
+        // SECURITY (R240-P3-PROXY-5): Compare in u64 space to avoid truncation on 32-bit.
         if let Some(len) = resp.content_length() {
-            if len as usize > MAX_JWKS_BODY_BYTES {
+            if len > MAX_JWKS_BODY_BYTES as u64 {
                 return Err(FederationError::JwksFetchFailed {
                     org_id: org_id.to_string(),
                     source: format!(
