+++ vellaveto-http-proxy/src/proxy/fallback.rs
@@ -112,8 +112,9 @@ pub async fn forward_with_fallback(
                 let status = resp.status().as_u16();
 
                 // SECURITY (FIND-R42-020): Fast-reject if Content-Length exceeds limit.
+                // SECURITY (R240-P3-PROXY-3): Compare in u64 space to avoid truncation on 32-bit.
                 if let Some(len) = resp.content_length() {
-                    if len as usize > MAX_RESPONSE_BODY_BYTES {
+                    if len > MAX_RESPONSE_BODY_BYTES as u64 {
                         last_error = format!(
                             "response body too large: {len} bytes (max {MAX_RESPONSE_BODY_BYTES})"
                         );
